+++ Season_Shift/Season_Shift/ParticleSystemComponent.h
@@ -17,7 +17,7 @@ private:
 		EmitterMetaData(float lifeTime, float particlesPerSecond, DirectX::SimpleMath::Vector3 offset);
 	};
 	unsigned int m_maxParticles;
-	unsigned int m_particleLifeTime;
+	float m_particleLifeTime;
 	std::shared_ptr<ParticleSystem> m_partSys;
 	std::string m_simShaderFile;
 	std::string m_emittShaderFile;
@@ -29,8 +29,8 @@ private:
 	void clearComponent() override;
 	void update() override;
 public:
-	ParticleSystemComponent(unsigned int maxParticles, unsigned int particleLifeTime);
-	ParticleSystemComponent(const std::string& simShader, const std::string& emittShader, unsigned int maxParticles, unsigned int particleLifeTime);
+	ParticleSystemComponent(unsigned int maxParticles, float particleLifeTime);
+	ParticleSystemComponent(const std::string& simShader, const std::string& emittShader, unsigned int maxParticles, float particleLifeTime);
 	int addEmitter(float particlesPerSecond, float emitterLifeTime, float scale, DirectX::SimpleMath::Vector3 color, DirectX::SimpleMath::Vector3 offset = DirectX::SimpleMath::Vector3::Zero,
 		float particleStartLifeTime = 0, DirectX::SimpleMath::Vector3 direction = DirectX::SimpleMath::Vector3::Zero);
 	void stopEmitter(int id);
